+++ utils/file_times_linux.go
@@ -10,5 +10,6 @@ import (
 // getBirthTime returns the change time (ctime) on Linux as a fallback,
 // since true birth time (btime) is not universally available across all filesystems.
 func getBirthTime(stat *syscall.Stat_t) time.Time {
-	return time.Unix(stat.Ctim.Sec, stat.Ctim.Nsec)
+	// Cast to int64 for 32-bit compatibility where Ctim fields are int32
+	return time.Unix(int64(stat.Ctim.Sec), int64(stat.Ctim.Nsec))
 }
